Move usage components to folder
